remove old kvClient and puller
