+++ cdc/kv/shared_client.go
@@ -16,6 +16,7 @@ package kv
 import (
 	"context"
 	"encoding/binary"
+	"fmt"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -46,6 +47,116 @@ import (
 	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
 )
+
+const (
+	dialTimeout           = 10 * time.Second
+	tikvRequestMaxBackoff = 20000 // Maximum total sleep time(in ms)
+
+	// TiCDC may open numerous gRPC streams,
+	// with 65535 bytes window size, 10K streams takes about 27GB memory.
+	//
+	// 65535 bytes, the initial window size in http2 spec.
+	grpcInitialWindowSize = (1 << 16) - 1
+	// 8 MB The value for initial window size on a connection
+	grpcInitialConnWindowSize = 1 << 23
+	// 256 MB The maximum message size the client can receive
+	grpcMaxCallRecvMsgSize = 1 << 28
+
+	// The threshold of warning a message is too large. TiKV split events into 6MB per-message.
+	warnRecvMsgSizeThreshold = 12 * 1024 * 1024
+
+	// TiCDC always interacts with region leader, every time something goes wrong,
+	// failed region will be reloaded via `BatchLoadRegionsWithKeyRange` API. So we
+	// don't need to force reload region anymore.
+	regionScheduleReload = false
+
+	scanRegionsConcurrency = 1024
+
+	tableMonitorInterval = 2 * time.Second
+)
+
+// time interval to force kv client to terminate gRPC stream and reconnect
+var reconnectInterval = 60 * time.Minute
+
+// streamAlterInterval is the interval to limit the frequency of creating/deleting streams.
+// Make it a variable so that we can change it in unit test.
+var streamAlterInterval = 1 * time.Second
+
+type regionStatefulEvent struct {
+	changeEvent     *cdcpb.Event
+	resolvedTsEvent *resolvedTsEvent
+	state           *regionFeedState
+
+	// regionID is used for load balancer, we don't use fields in state to reduce lock usage
+	regionID uint64
+
+	// finishedCallbackCh is used to mark events that are sent from a give region
+	// worker to this worker(one of the workers in worker pool) are all processed.
+	finishedCallbackCh chan struct{}
+}
+
+type resolvedTsEvent struct {
+	resolvedTs uint64
+	regions    []*regionFeedState
+}
+
+var (
+	metricFeedNotLeaderCounter        = eventFeedErrorCounter.WithLabelValues("NotLeader")
+	metricFeedEpochNotMatchCounter    = eventFeedErrorCounter.WithLabelValues("EpochNotMatch")
+	metricFeedRegionNotFoundCounter   = eventFeedErrorCounter.WithLabelValues("RegionNotFound")
+	metricFeedDuplicateRequestCounter = eventFeedErrorCounter.WithLabelValues("DuplicateRequest")
+	metricFeedUnknownErrorCounter     = eventFeedErrorCounter.WithLabelValues("Unknown")
+	metricFeedRPCCtxUnavailable       = eventFeedErrorCounter.WithLabelValues("RPCCtxUnavailable")
+	metricStoreSendRequestErr         = eventFeedErrorCounter.WithLabelValues("SendRequestToStore")
+	metricConnectToStoreErr           = eventFeedErrorCounter.WithLabelValues("ConnectToStore")
+	metricKvIsBusyCounter             = eventFeedErrorCounter.WithLabelValues("KvIsBusy")
+)
+
+var (
+	// unreachable error, only used in unit test
+	errUnreachable = errors.New("kv client unreachable error")
+	// internal error, force the gPRC stream terminate and reconnect
+	errReconnect = errors.New("internal error, reconnect all regions")
+	logPanic     = log.Panic
+)
+
+type regionErrorInfo struct {
+	singleRegionInfo
+	err error
+}
+
+func newRegionErrorInfo(info singleRegionInfo, err error) regionErrorInfo {
+	return regionErrorInfo{
+		singleRegionInfo: info,
+		err:              err,
+	}
+}
+
+type eventError struct {
+	err *cdcpb.Error
+}
+
+// Error implement error interface.
+func (e *eventError) Error() string {
+	return e.err.String()
+}
+
+type rpcCtxUnavailableErr struct {
+	verID tikv.RegionVerID
+}
+
+func (e *rpcCtxUnavailableErr) Error() string {
+	return fmt.Sprintf("cannot get rpcCtx for region %v. ver:%v, confver:%v",
+		e.verID.GetID(), e.verID.GetVer(), e.verID.GetConfVer())
+}
+
+type connectToStoreErr struct{}
+
+func (e *connectToStoreErr) Error() string { return "connect to store error" }
+
+type sendRequestToStoreErr struct{}
+
+func (e *sendRequestToStoreErr) Error() string { return "send request to store error" }
 
 // SubscriptionID comes from `SharedClient.AllocSubscriptionID`.
 type SubscriptionID uint64
