+++ cdc/processor/sourcemanager/manager.go
@@ -28,7 +28,6 @@ import (
 	"github.com/pingcap/tiflow/cdc/processor/tablepb"
 	"github.com/pingcap/tiflow/cdc/puller"
 	"github.com/pingcap/tiflow/pkg/config"
-	"github.com/pingcap/tiflow/pkg/spanz"
 	"github.com/pingcap/tiflow/pkg/txnutil"
 	"github.com/pingcap/tiflow/pkg/upstream"
 	"github.com/tikv/client-go/v2/tikv"
@@ -36,21 +35,6 @@ import (
 )
 
 const defaultMaxBatchSize = 256
-
-type pullerWrapperCreator func(
-	changefeed model.ChangeFeedID,
-	span tablepb.Span,
-	tableName string,
-	startTs model.Ts,
-	bdrMode bool,
-) pullerwrapper.Wrapper
-
-type tablePullers struct {
-	ctx     context.Context
-	errChan chan error
-	spanz.SyncMap
-	pullerWrapperCreator pullerWrapperCreator
-}
 
 type multiplexingPuller struct {
 	puller *pullerwrapper.MultiplexingWrapper
@@ -72,12 +56,7 @@ type SourceManager struct {
 	// Used to indicate whether the changefeed is in BDR mode.
 	bdrMode bool
 
-	// if `multiplexing` is true (the default value) then `multiplexingPuller` will be used.
-	//  * tables in one changefeed will share grpc streams and region workers
-	//  * tables in one changefeed will share goroutines
-	multiplexing       bool
 	enableTableMonitor bool
-	tablePullers       tablePullers
 	multiplexingPuller multiplexingPuller
 }
 
@@ -90,8 +69,7 @@ func New(
 	bdrMode bool,
 	enableTableMonitor bool,
 ) *SourceManager {
-	multiplexing := config.GetGlobalServerConfig().KVClient.EnableMultiplexing
-	return newSourceManager(changefeedID, up, mg, engine, bdrMode, multiplexing, pullerwrapper.NewPullerWrapper, enableTableMonitor)
+	return newSourceManager(changefeedID, up, mg, engine, bdrMode, enableTableMonitor)
 }
 
 // NewForTest creates a new source manager for testing.
@@ -102,7 +80,7 @@ func NewForTest(
 	engine sorter.SortEngine,
 	bdrMode bool,
 ) *SourceManager {
-	return newSourceManager(changefeedID, up, mg, engine, bdrMode, false, pullerwrapper.NewPullerWrapperForTest, false)
+	return newSourceManager(changefeedID, up, mg, engine, bdrMode, false)
 }
 
 func newSourceManager(
@@ -111,8 +89,6 @@ func newSourceManager(
 	mg entry.MounterGroup,
 	engine sorter.SortEngine,
 	bdrMode bool,
-	multiplexing bool,
-	pullerWrapperCreator pullerWrapperCreator,
 	enableTableMonitor bool,
 ) *SourceManager {
 	mgr := &SourceManager{
@@ -122,12 +98,7 @@ func newSourceManager(
 		mg:                 mg,
 		engine:             engine,
 		bdrMode:            bdrMode,
-		multiplexing:       multiplexing,
 		enableTableMonitor: enableTableMonitor,
-	}
-	if !multiplexing {
-		mgr.tablePullers.errChan = make(chan error, 16)
-		mgr.tablePullers.pullerWrapperCreator = pullerWrapperCreator
 	}
 	return mgr
 }
@@ -137,27 +108,12 @@ func (m *SourceManager) AddTable(span tablepb.Span, tableName string, startTs mo
 	// Add table to the engine first, so that the engine can receive the events from the puller.
 	m.engine.AddTable(span, startTs)
 
-	if m.multiplexing {
-		m.multiplexingPuller.puller.Subscribe([]tablepb.Span{span}, startTs, tableName)
-		return
-	}
-
-	p := m.tablePullers.pullerWrapperCreator(m.changefeedID, span, tableName, startTs, m.bdrMode)
-	p.Start(m.tablePullers.ctx, m.up, m.engine, m.tablePullers.errChan, m.enableTableMonitor)
-	m.tablePullers.Store(span, p)
+	m.multiplexingPuller.puller.Subscribe([]tablepb.Span{span}, startTs, tableName)
 }
 
 // RemoveTable removes a table from the source manager. Stop puller and unregister table from the engine.
 func (m *SourceManager) RemoveTable(span tablepb.Span) {
-	if m.multiplexing {
-		m.multiplexingPuller.puller.Unsubscribe([]tablepb.Span{span})
-		m.engine.RemoveTable(span)
-		return
-	}
-
-	if wrapper, ok := m.tablePullers.LoadAndDelete(span); ok {
-		wrapper.(pullerwrapper.Wrapper).Close()
-	}
+	m.multiplexingPuller.puller.Unsubscribe([]tablepb.Span{span})
 	m.engine.RemoveTable(span)
 }
 
@@ -182,18 +138,7 @@ func (m *SourceManager) CleanByTable(span tablepb.Span, upperBound sorter.Positi
 
 // GetTablePullerStats returns the puller stats of the table.
 func (m *SourceManager) GetTablePullerStats(span tablepb.Span) puller.Stats {
-	if m.multiplexing {
-		return m.multiplexingPuller.puller.MultiplexingPuller.Stats(span)
-	}
-
-	p, ok := m.tablePullers.Load(span)
-	if !ok {
-		log.Panic("Table puller not found when getting table puller stats",
-			zap.String("namespace", m.changefeedID.Namespace),
-			zap.String("changefeed", m.changefeedID.ID),
-			zap.Stringer("span", &span))
-	}
-	return p.(pullerwrapper.Wrapper).GetStats()
+	return m.multiplexingPuller.puller.MultiplexingPuller.Stats(span)
 }
 
 // GetTableSorterStats returns the sorter stats of the table.
@@ -203,31 +148,20 @@ func (m *SourceManager) GetTableSorterStats(span tablepb.Span) sorter.TableStats
 
 // Run implements util.Runnable.
 func (m *SourceManager) Run(ctx context.Context, _ ...chan<- error) error {
-	if m.multiplexing {
-		serverConfig := config.GetGlobalServerConfig()
-		grpcPool := sharedconn.NewConnAndClientPool(m.up.SecurityConfig, kv.GetGlobalGrpcMetrics())
-		client := kv.NewSharedClient(
-			m.changefeedID, serverConfig, m.bdrMode,
-			m.up.PDClient, grpcPool, m.up.RegionCache, m.up.PDClock,
-			txnutil.NewLockerResolver(m.up.KVStorage.(tikv.Storage), m.changefeedID),
-		)
-		m.multiplexingPuller.puller = pullerwrapper.NewMultiplexingPullerWrapper(
-			m.changefeedID, client, m.engine,
-			int(serverConfig.KVClient.FrontierConcurrent),
-		)
+	serverConfig := config.GetGlobalServerConfig()
+	grpcPool := sharedconn.NewConnAndClientPool(m.up.SecurityConfig, kv.GetGlobalGrpcMetrics())
+	client := kv.NewSharedClient(
+		m.changefeedID, serverConfig, m.bdrMode,
+		m.up.PDClient, grpcPool, m.up.RegionCache, m.up.PDClock,
+		txnutil.NewLockerResolver(m.up.KVStorage.(tikv.Storage), m.changefeedID),
+	)
+	m.multiplexingPuller.puller = pullerwrapper.NewMultiplexingPullerWrapper(
+		m.changefeedID, client, m.engine,
+		int(serverConfig.KVClient.FrontierConcurrent),
+	)
 
-		close(m.ready)
-		return m.multiplexingPuller.puller.Run(ctx)
-	}
-
-	m.tablePullers.ctx = ctx
 	close(m.ready)
-	select {
-	case err := <-m.tablePullers.errChan:
-		return err
-	case <-m.tablePullers.ctx.Done():
-		return m.tablePullers.ctx.Err()
-	}
+	return m.multiplexingPuller.puller.Run(ctx)
 }
 
 // WaitForReady implements util.Runnable.
@@ -246,10 +180,7 @@ func (m *SourceManager) Close() {
 		zap.String("changefeed", m.changefeedID.ID))
 
 	start := time.Now()
-	m.tablePullers.Range(func(span tablepb.Span, value interface{}) bool {
-		value.(pullerwrapper.Wrapper).Close()
-		return true
-	})
+
 	log.Info("All pullers have been closed",
 		zap.String("namespace", m.changefeedID.Namespace),
 		zap.String("changefeed", m.changefeedID.ID),
