+++ cdc/server/server.go
@@ -31,7 +31,6 @@ import (
 	"github.com/pingcap/tidb/pkg/util/gctuner"
 	"github.com/pingcap/tiflow/cdc"
 	"github.com/pingcap/tiflow/cdc/capture"
-	"github.com/pingcap/tiflow/cdc/kv"
 	"github.com/pingcap/tiflow/cdc/processor/sourcemanager/sorter/factory"
 	capturev2 "github.com/pingcap/tiflow/cdcv2/capture"
 	"github.com/pingcap/tiflow/pkg/config"
@@ -345,10 +344,6 @@ func (s *server) run(ctx context.Context) (err error) {
 
 	eg.Go(func() error {
 		return s.upstreamPDHealthChecker(egCtx)
-	})
-
-	eg.Go(func() error {
-		return kv.RunWorkerPool(egCtx)
 	})
 
 	eg.Go(func() error {
